Move Ktor config to config
