+++ src/main/kotlin/io/github/nomisrev/config/ktor.kt
@@ -1,0 +1,32 @@
+package io.github.nomisrev.config
+
+import io.ktor.http.HttpHeaders
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.CORS
+import io.ktor.server.plugins.ContentNegotiation
+import io.ktor.server.plugins.DefaultHeaders
+import io.ktor.server.plugins.maxAgeDuration
+import kotlin.time.Duration.Companion.days
+import kotlin.time.ExperimentalTime
+import kotlinx.serialization.json.Json
+
+@OptIn(ExperimentalTime::class)
+fun Application.configure() {
+  install(DefaultHeaders)
+  install(ContentNegotiation) {
+    json(
+      Json {
+        isLenient = true
+        ignoreUnknownKeys = true
+      }
+    )
+  }
+  install(CORS) {
+    header(HttpHeaders.Authorization)
+    header(HttpHeaders.ContentType)
+    allowNonSimpleContentTypes = true
+    maxAgeDuration = 3.days
+  }
+}
